driving temp to control eforce "works"
Works in the sense that temperature is adjusted until the extruder force
stabilizes at the desired value. At least if the extrusion rate is more
or less constant over several minutes.

Doesn't work in the sense that it is not of much use when a print slows
down for cooling.  During slow-down the e-force will drop, causing the
new code to drop the temperature down to 160C (nozzle clogged) long before
the e-force starts to increase again. At which point the new code will
increase the temperature and un-clog, but it's much too late.